Separate vdb reader from index
