--- src/openvdb/core/OpenVDB.js
@@ -1,4 +1,0 @@
-
-class OpenVDB {
-
-}
